compact condition build files
